Move service sorting to config package.
